MAINT: pipefunc.map module reorganization (#397)
