+++ pipefunc/map/_run.py
@@ -4,103 +4,39 @@ import asyncio
 import functools
 import itertools
 import time
-import warnings
-from collections import OrderedDict, defaultdict
-from concurrent.futures import Executor, Future, ProcessPoolExecutor, ThreadPoolExecutor
+from concurrent.futures import Executor, Future, ProcessPoolExecutor
 from contextlib import contextmanager
-from dataclasses import dataclass
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, NamedTuple, TypeAlias
 
 import numpy as np
 import numpy.typing as npt
 
-from pipefunc._utils import (
-    at_least_tuple,
-    dump,
-    handle_error,
-    is_running_in_ipynb,
-    load,
-    prod,
-    requires,
-)
+from pipefunc._utils import at_least_tuple, dump, handle_error, is_running_in_ipynb, load, prod
 from pipefunc.cache import HybridCache, to_hashable
-from pipefunc.map._mapspec import MapSpec, _shape_to_key, validate_consistent_axes
-from pipefunc.map._run_info import DirectValue, RunInfo, _external_shape, _internal_shape
-from pipefunc.map._storage._base import StorageBase, _iterate_shape_indices, _select_by_mask
+from pipefunc.map._mapspec import MapSpec, _shape_to_key
+from pipefunc.map._storage_array._base import StorageBase, iterate_shape_indices, select_by_mask
+
+from ._prepare import prepare_run
+from ._result import DirectValue, Result
+from ._shapes import external_shape_from_mask, internal_shape_from_mask
 
 if TYPE_CHECKING:
+    from collections import OrderedDict
     from collections.abc import Callable, Generator, Sequence
-
-    import xarray as xr
 
     from pipefunc import PipeFunc, Pipeline
     from pipefunc._widgets import ProgressTracker
     from pipefunc.cache import _CacheBase
 
+    from ._progress import Status
+    from ._run_info import RunInfo
+
 
 _OUTPUT_TYPE: TypeAlias = str | tuple[str, ...]
 
 
-def _cannot_be_parallelized(pipeline: Pipeline) -> bool:
-    return all(f.mapspec is None for f in pipeline.functions) and all(
-        len(fs) == 1 for fs in pipeline.topological_generations.function_lists
-    )
-
-
-def _prepare_run(
-    *,
-    pipeline: Pipeline,
-    inputs: dict[str, Any],
-    run_folder: str | Path | None,
-    internal_shapes: dict[str, int | tuple[int, ...]] | None,
-    output_names: set[_OUTPUT_TYPE] | None,
-    parallel: bool,
-    executor: Executor | dict[_OUTPUT_TYPE, Executor] | None,
-    storage: str | dict[_OUTPUT_TYPE, str],
-    cleanup: bool,
-    fixed_indices: dict[str, int | slice] | None,
-    auto_subpipeline: bool,
-    show_progress: bool,
-    in_async: bool,
-) -> tuple[
-    Pipeline,
-    RunInfo,
-    dict[str, StorageBase | Path | DirectValue],
-    OrderedDict[str, Result],
-    bool,
-    ProgressTracker | None,
-]:
-    if not parallel and show_progress:
-        msg = "Cannot use `show_progress=True` with `parallel=False`."
-        raise ValueError(msg)
-    if not parallel and executor:
-        msg = "Cannot use an executor without `parallel=True`."
-        raise ValueError(msg)
-    inputs = pipeline._flatten_scopes(inputs)
-    if auto_subpipeline or output_names is not None:
-        pipeline = pipeline.subpipeline(set(inputs), output_names)
-    _validate_complete_inputs(pipeline, inputs)
-    validate_consistent_axes(pipeline.mapspecs(ordered=False))
-    _validate_fixed_indices(fixed_indices, inputs, pipeline)
-    run_info = RunInfo.create(
-        run_folder,
-        pipeline,
-        inputs,
-        internal_shapes,
-        storage=storage,
-        cleanup=cleanup,
-    )
-    outputs: OrderedDict[str, Result] = OrderedDict()
-    store = run_info.init_store()
-    progress = _init_tracker(store, pipeline.sorted_functions, show_progress, in_async)
-    if executor is None and _cannot_be_parallelized(pipeline):
-        parallel = False
-    _check_parallel(parallel, store, executor)
-    return pipeline, run_info, store, outputs, parallel, progress
-
-
-def run(
+def run_map(
     pipeline: Pipeline,
     inputs: dict[str, Any],
     run_folder: str | Path | None = None,
@@ -179,7 +115,7 @@ def run(
         Whether to display a progress bar. Only works if ``parallel=True``.
 
     """
-    pipeline, run_info, store, outputs, parallel, progress = _prepare_run(
+    pipeline, run_info, store, outputs, parallel, progress = prepare_run(
         pipeline=pipeline,
         inputs=inputs,
         run_folder=run_folder,
@@ -214,7 +150,7 @@ def run(
     return outputs
 
 
-class AsyncRun(NamedTuple):
+class AsyncMap(NamedTuple):
     task: asyncio.Task[OrderedDict[str, Result]]
     run_info: RunInfo
     progress: ProgressTracker | None
@@ -233,7 +169,7 @@ class AsyncRun(NamedTuple):
         return loop.run_until_complete(self.task)  # pragma: no cover
 
 
-def run_async(
+def run_map_async(
     pipeline: Pipeline,
     inputs: dict[str, Any],
     run_folder: str | Path | None = None,
@@ -247,7 +183,7 @@ def run_async(
     fixed_indices: dict[str, int | slice] | None = None,
     auto_subpipeline: bool = False,
     show_progress: bool = False,
-) -> AsyncRun:
+) -> AsyncMap:
     """Asynchronously run a pipeline with `MapSpec` functions for given ``inputs``.
 
     Returns immediately with an `AsyncRun` instance with a `task` attribute that can be awaited.
@@ -309,7 +245,7 @@ def run_async(
         Whether to display a progress bar.
 
     """
-    pipeline, run_info, store, outputs, _, progress = _prepare_run(
+    pipeline, run_info, store, outputs, _, progress = prepare_run(
         pipeline=pipeline,
         inputs=inputs,
         run_folder=run_folder,
@@ -346,7 +282,7 @@ def run_async(
     if progress is not None:
         progress.attach_task(task)
         progress.display()
-    return AsyncRun(task, run_info, progress)
+    return AsyncMap(task, run_info, progress)
 
 
 def _maybe_persist_memory(
@@ -357,58 +293,6 @@ def _maybe_persist_memory(
         for arr in store.values():
             if isinstance(arr, StorageBase):
                 arr.persist()
-
-
-class Result(NamedTuple):
-    function: str
-    kwargs: dict[str, Any]
-    output_name: str
-    output: Any
-    store: StorageBase | Path | DirectValue
-
-
-def load_outputs(*output_names: str, run_folder: str | Path) -> Any:
-    """Load the outputs of a run."""
-    run_folder = Path(run_folder)
-    run_info = RunInfo.load(run_folder)
-    store = run_info.init_store()
-    outputs = [_load_from_store(output_name, store).value for output_name in output_names]
-    outputs = [_maybe_load_array(o) for o in outputs]
-    return outputs[0] if len(output_names) == 1 else outputs
-
-
-def load_xarray_dataset(
-    *output_name: str,
-    run_folder: str | Path,
-    load_intermediate: bool = True,
-) -> xr.Dataset:
-    """Load the output(s) of a `pipeline.map` as an `xarray.Dataset`.
-
-    Parameters
-    ----------
-    output_name
-        The names of the outputs to load. If empty, all outputs are loaded.
-    run_folder
-        The folder where the pipeline run was stored.
-    load_intermediate
-        Whether to load intermediate outputs as coordinates.
-
-    Returns
-    -------
-        An `xarray.Dataset` containing the outputs of the pipeline run.
-
-    """
-    requires("xarray", reason="load_xarray_dataset", extras="xarray")
-    from pipefunc.map.xarray import load_xarray_dataset
-
-    run_info = RunInfo.load(run_folder)
-    return load_xarray_dataset(
-        run_info.mapspecs,
-        run_info.inputs,
-        run_folder=run_folder,
-        output_names=output_name,  # type: ignore[arg-type]
-        load_intermediate=load_intermediate,
-    )
 
 
 def _dump_single_output(
@@ -473,11 +357,11 @@ def _select_kwargs(
     index: int,
 ) -> dict[str, Any]:
     assert func.mapspec is not None
-    external_shape = _external_shape(shape, shape_mask)
+    external_shape = external_shape_from_mask(shape, shape_mask)
     input_keys = func.mapspec.input_keys(external_shape, index)
     normalized_keys = {k: v[0] if len(v) == 1 else v for k, v in input_keys.items()}
     selected = {k: v[normalized_keys[k]] if k in normalized_keys else v for k, v in kwargs.items()}
-    _load_file_arrays(selected)
+    _load_arrays(selected)
     return selected
 
 
@@ -546,28 +430,28 @@ def _run_iteration_and_process(
     kwargs: dict[str, Any],
     shape: tuple[int, ...],
     shape_mask: tuple[bool, ...],
-    file_arrays: Sequence[StorageBase],
+    arrays: Sequence[StorageBase],
     cache: _CacheBase | None = None,
 ) -> tuple[Any, ...]:
     output = _run_iteration(func, kwargs, shape, shape_mask, index, cache)
     outputs = _pick_output(func, output)
-    _update_file_array(func, file_arrays, shape, shape_mask, index, outputs)
+    _update_array(func, arrays, shape, shape_mask, index, outputs)
     return outputs
 
 
-def _update_file_array(
+def _update_array(
     func: PipeFunc,
-    file_arrays: Sequence[StorageBase],
+    arrays: Sequence[StorageBase],
     shape: tuple[int, ...],
     shape_mask: tuple[bool, ...],
     index: int,
     outputs: tuple[Any, ...],
 ) -> None:
     assert isinstance(func.mapspec, MapSpec)
-    external_shape = _external_shape(shape, shape_mask)
+    external_shape = external_shape_from_mask(shape, shape_mask)
     output_key = func.mapspec.output_key(external_shape, index)
-    for file_array, _output in zip(file_arrays, outputs):
-        file_array.dump(output_key, _output)
+    for array, _output in zip(arrays, outputs):
+        array.dump(output_key, _output)
 
 
 def _indices_to_flat_index(
@@ -577,8 +461,8 @@ def _indices_to_flat_index(
     external_index: tuple[int, ...],
     internal_index: tuple[int, ...],
 ) -> np.int_:
-    full_index = _select_by_mask(shape_mask, external_index, internal_index)
-    full_shape = _select_by_mask(shape_mask, shape, internal_shape)
+    full_index = select_by_mask(shape_mask, external_index, internal_index)
+    full_shape = select_by_mask(shape_mask, shape, internal_shape)
     return np.ravel_multi_index(full_index, full_shape)
 
 
@@ -589,11 +473,11 @@ def _set_output(
     shape: tuple[int, ...],
     shape_mask: tuple[bool, ...],
 ) -> None:
-    external_shape = _external_shape(shape, shape_mask)
-    internal_shape = _internal_shape(shape, shape_mask)
+    external_shape = external_shape_from_mask(shape, shape_mask)
+    internal_shape = internal_shape_from_mask(shape, shape_mask)
     external_index = _shape_to_key(external_shape, linear_index)
     assert np.shape(output) == internal_shape
-    for internal_index in _iterate_shape_indices(internal_shape):
+    for internal_index in iterate_shape_indices(internal_shape):
         flat_index = _indices_to_flat_index(
             external_shape,
             internal_shape,
@@ -620,12 +504,12 @@ def _update_result_array(
 
 
 def _existing_and_missing_indices(
-    file_arrays: list[StorageBase],
+    arrays: list[StorageBase],
     fixed_mask: np.flatiter[npt.NDArray[np.bool_]] | None,
 ) -> tuple[list[int], list[int]]:
     # TODO: when `fixed_indices` are used we could be more efficient by not
     # computing the full mask.
-    masks = (arr.mask_linear() for arr in file_arrays)
+    masks = (arr.mask_linear() for arr in arrays)
     if fixed_mask is None:
         fixed_mask = itertools.repeat(object=True)  # type: ignore[assignment]
 
@@ -660,7 +544,7 @@ class _MapSpecArgs(NamedTuple):
     result_arrays: list[np.ndarray]
     shape: tuple[int, ...]
     mask: tuple[bool, ...]
-    file_arrays: list[StorageBase]
+    arrays: list[StorageBase]
 
 
 def _prepare_submit_map_spec(
@@ -674,7 +558,7 @@ def _prepare_submit_map_spec(
     assert isinstance(func.mapspec, MapSpec)
     shape = run_info.shapes[func.output_name]
     mask = run_info.shape_masks[func.output_name]
-    file_arrays: list[StorageBase] = [store[name] for name in at_least_tuple(func.output_name)]  # type: ignore[misc]
+    arrays: list[StorageBase] = [store[name] for name in at_least_tuple(func.output_name)]  # type: ignore[misc]
     result_arrays = _init_result_arrays(func.output_name, shape)
     process_index = functools.partial(
         _run_iteration_and_process,
@@ -682,12 +566,12 @@ def _prepare_submit_map_spec(
         kwargs=kwargs,
         shape=shape,
         shape_mask=mask,
-        file_arrays=file_arrays,
+        arrays=arrays,
         cache=cache,
     )
     fixed_mask = _mask_fixed_axes(fixed_indices, func.mapspec, shape, mask)
-    existing, missing = _existing_and_missing_indices(file_arrays, fixed_mask)  # type: ignore[arg-type]
-    return _MapSpecArgs(process_index, existing, missing, result_arrays, shape, mask, file_arrays)
+    existing, missing = _existing_and_missing_indices(arrays, fixed_mask)  # type: ignore[arg-type]
+    return _MapSpecArgs(process_index, existing, missing, result_arrays, shape, mask, arrays)
 
 
 def _mask_fixed_axes(
@@ -699,8 +583,8 @@ def _mask_fixed_axes(
     if fixed_indices is None:
         return None
     key = tuple(fixed_indices.get(axis, slice(None)) for axis in mapspec.output_indices)
-    external_key = _external_shape(key, shape_mask)  # type: ignore[arg-type]
-    external_shape = _external_shape(shape, shape_mask)
+    external_key = external_shape_from_mask(key, shape_mask)  # type: ignore[arg-type]
+    external_shape = external_shape_from_mask(shape, shape_mask)
     select: npt.NDArray[np.bool_] = np.zeros(external_shape, dtype=bool)
     select[external_key] = True
     return select.flat
@@ -709,7 +593,7 @@ def _mask_fixed_axes(
 def _status_submit(
     func: Callable[..., Any],
     executor: Executor,
-    status: _Status,
+    status: Status,
     progress: ProgressTracker,
     *args: Any,
 ) -> Future:
@@ -725,7 +609,7 @@ def _maybe_parallel_map(
     func: Callable[..., Any],
     seq: Sequence,
     executor: Executor | None,
-    status: _Status | None,
+    status: Status | None,
     progress: ProgressTracker | None,
 ) -> list[Any]:
     if executor is not None:
@@ -739,7 +623,7 @@ def _maybe_parallel_map(
 def _maybe_submit(
     func: Callable[..., Any],
     executor: Executor | None,
-    status: _Status | None,
+    status: Status | None,
     progress: ProgressTracker | None,
     *args: Any,
 ) -> Any:
@@ -803,7 +687,7 @@ def _submit_single(
         return output
 
     # Otherwise, run the function
-    _load_file_arrays(kwargs)
+    _load_arrays(kwargs)
 
     def compute_fn() -> Any:
         try:
@@ -822,7 +706,7 @@ def _maybe_load_array(x: Any) -> Any:
     return x
 
 
-def _load_file_arrays(kwargs: dict[str, Any]) -> None:
+def _load_arrays(kwargs: dict[str, Any]) -> None:
     for k, v in kwargs.items():
         kwargs[k] = _maybe_load_array(v)
 
@@ -830,63 +714,6 @@ def _load_file_arrays(kwargs: dict[str, Any]) -> None:
 class _KwargsTask(NamedTuple):
     kwargs: dict[str, Any]
     task: tuple[Any, _MapSpecArgs] | Any
-
-
-@dataclass
-class _Status:
-    """A class to keep track of the progress of a function."""
-
-    n_total: int
-    n_in_progress: int = 0
-    n_completed: int = 0
-    n_failed: int = 0
-    start_time: float | None = None
-    end_time: float | None = None
-
-    @property
-    def n_left(self) -> int:
-        return self.n_total - self.n_completed - self.n_failed
-
-    def mark_in_progress(self) -> None:
-        if self.start_time is None:
-            self.start_time = time.monotonic()
-        self.n_in_progress += 1
-
-    def mark_complete(self, _: Any) -> None:  # needs arg to be used as callback
-        self.n_in_progress -= 1
-        self.n_completed += 1
-        if self.n_completed == self.n_total:
-            self.end_time = time.monotonic()
-
-    @property
-    def progress(self) -> float:
-        return self.n_completed / self.n_total
-
-    def elapsed_time(self) -> float:
-        assert self.start_time is not None
-        if self.end_time is None:
-            return time.monotonic() - self.start_time
-        return self.end_time - self.start_time
-
-
-def _init_tracker(
-    store: dict[str, StorageBase | Path | DirectValue],
-    functions: list[PipeFunc],
-    show_progress: bool,  # noqa: FBT001
-    in_async: bool,  # noqa: FBT001
-) -> ProgressTracker | None:
-    if not show_progress:
-        return None
-    requires("ipywidgets", reason="show_progress", extras="ipywidgets")
-    from pipefunc._widgets import ProgressTracker
-
-    progress = {}
-    for func in functions:
-        name, *_ = at_least_tuple(func.output_name)  # if multiple, the have equal size
-        s = store[name]
-        size = s.size if isinstance(s, StorageBase) else 1
-        progress[func.output_name] = _Status(n_total=size)
-    return ProgressTracker(progress, None, display=False, in_async=in_async)
 
 
 # NOTE: A similar async version of this function is provided below.
@@ -1020,7 +847,7 @@ def _output_from_mapspec_task(
         _update_result_array(args.result_arrays, index, outputs, args.shape, args.mask)
 
     for index in args.existing:
-        outputs = [file_array.get_from_index(index) for file_array in args.file_arrays]
+        outputs = [array.get_from_index(index) for array in args.arrays]
         _update_result_array(args.result_arrays, index, outputs, args.shape, args.mask)
 
     return tuple(x.reshape(args.shape) for x in args.result_arrays)
@@ -1088,134 +915,3 @@ async def _process_task_async(
         r = await _result_async(task, loop)
         output = _dump_single_output(func, r, store)
     return _to_result_dict(func, kwargs, output, store)
-
-
-def _check_parallel(
-    parallel: bool,  # noqa: FBT001
-    store: dict[str, StorageBase | Path | DirectValue],
-    executor: Executor | dict[_OUTPUT_TYPE, Executor] | None,
-) -> None:
-    if isinstance(executor, dict):
-        uses_default_executor: set[str] = set(store.keys()) - {
-            n for name in executor for n in at_least_tuple(name)
-        }
-        for output_name, ex in executor.items():
-            names = uses_default_executor if output_name == "" else at_least_tuple(output_name)
-            _check_parallel(parallel, {n: store[n] for n in names}, ex)
-        return
-    if isinstance(executor, ThreadPoolExecutor):
-        return
-    if not parallel or not store:
-        return
-    for storage in store.values():
-        if isinstance(storage, StorageBase) and not storage.parallelizable:
-            recommendation = (
-                "Consider\n - using a file-based storage or `shared_memory` / `zarr_shared_memory`"
-                " for parallel execution,\n - disable parallel execution,\n - or use a different executor.\n"
-            )
-            default = f"The chosen storage type `{storage.storage_id}` does not support process-based parallel execution."
-            if executor is None:
-                msg = (
-                    f"{default}"
-                    f" PipeFunc defaults to using a `ProcessPoolExecutor`, which requires a parallelizable storage."
-                    f" {recommendation}"
-                )
-                raise ValueError(msg)
-            assert executor is not None
-            msg = (
-                f"{default}"
-                f" If the current executor of type `{type(executor).__name__}` is process-based, it is incompatible."
-                f" {recommendation}"
-            )
-            warnings.warn(msg, stacklevel=2)
-
-
-def _validate_complete_inputs(pipeline: Pipeline, inputs: dict[str, Any]) -> None:
-    """Validate that all required inputs are provided.
-
-    Note that `output_name is None` means that all outputs are required!
-    This is in contrast to some other functions, where ``None`` means that the
-    `pipeline.unique_leaf_node` is used.
-    """
-    root_args = set(pipeline.topological_generations.root_args)
-    inputs_with_defaults = set(inputs) | set(pipeline.defaults)
-    if missing := root_args - set(inputs_with_defaults):
-        missing_args = ", ".join(missing)
-        msg = f"Missing inputs: `{missing_args}`."
-        raise ValueError(msg)
-    if extra := set(inputs_with_defaults) - root_args:
-        extra_args = ", ".join(extra)
-        msg = f"Got extra inputs: `{extra_args}` that are not accepted by this pipeline."
-        raise ValueError(msg)
-
-
-def _validate_fixed_indices(
-    fixed_indices: dict[str, int | slice] | None,
-    inputs: dict[str, Any],
-    pipeline: Pipeline,
-) -> None:
-    if fixed_indices is None:
-        return
-    extra = set(fixed_indices)
-    axes = pipeline.mapspec_axes
-    for parameter, axes_ in axes.items():
-        for axis in axes_:
-            if axis in fixed_indices:
-                extra.discard(axis)
-        if parameter in inputs:
-            key = tuple(fixed_indices.get(axis, slice(None)) for axis in axes_)
-            if len(key) == 1:
-                key = key[0]  # type: ignore[assignment]
-            try:
-                inputs[parameter][key]
-            except IndexError as e:
-                msg = f"Fixed index `{key}` for parameter `{parameter}` is out of bounds."
-                raise IndexError(msg) from e
-    if extra:
-        msg = f"Got extra `fixed_indices`: `{extra}` that are not accepted by this map."
-        raise ValueError(msg)
-
-    reduced_axes = _reduced_axes(pipeline)
-    for name, axes_set in reduced_axes.items():
-        if reduced := set(axes_set) & set(fixed_indices):
-            reduced_str = ", ".join(reduced)
-            msg = f"Axis `{reduced_str}` in `{name}` is reduced and cannot be in `fixed_indices`."
-            raise ValueError(msg)
-
-
-def _reduced_axes(pipeline: Pipeline) -> dict[str, set[str]]:
-    # TODO: check the overlap between this an `independent_axes_in_mapspecs`.
-    # It might be that this function could be used instead.
-    reduced_axes: dict[str, set[str]] = defaultdict(set)
-    axes = pipeline.mapspec_axes
-    for name in pipeline.mapspec_names:
-        for func in pipeline.functions:
-            if _is_parameter_reduced_by_function(func, name):
-                reduced_axes[name].update(axes[name])
-            elif _is_parameter_partially_reduced_by_function(func, name):
-                _axes = _get_partially_reduced_axes(func, name, axes)
-                reduced_axes[name].update(_axes)
-    return dict(reduced_axes)
-
-
-def _is_parameter_reduced_by_function(func: PipeFunc, name: str) -> bool:
-    return name in func.parameters and (
-        func.mapspec is None or name not in func.mapspec.input_names
-    )
-
-
-def _is_parameter_partially_reduced_by_function(func: PipeFunc, name: str) -> bool:
-    if func.mapspec is None or name not in func.mapspec.input_names:
-        return False
-    spec = next(spec for spec in func.mapspec.inputs if spec.name == name)
-    return None in spec.axes
-
-
-def _get_partially_reduced_axes(
-    func: PipeFunc,
-    name: str,
-    axes: dict[str, tuple[str, ...]],
-) -> tuple[str, ...]:
-    assert func.mapspec is not None
-    spec = next(spec for spec in func.mapspec.inputs if spec.name == name)
-    return tuple(ax for ax, spec_ax in zip(axes[name], spec.axes) if spec_ax is None)
