+++ pipefunc/map/_storage_array/__init__.py
@@ -1,0 +1,1 @@
+"""pipefunc.map._storage: Modules that handle storage for MapSpecs."""
