+++ pipefunc/map/_storage_array/_zarr.py
@@ -15,7 +15,8 @@ from numcodecs.compat import ensure_contiguous_ndarray
 from numcodecs.registry import register_codec
 
 from pipefunc._utils import prod
-from pipefunc.map._storage._base import StorageBase, _select_by_mask, register_storage
+
+from ._base import StorageBase, register_storage, select_by_mask
 
 
 class ZarrFileArray(StorageBase):
@@ -52,7 +53,7 @@ class ZarrFileArray(StorageBase):
         if object_codec is None:
             object_codec = CloudPickleCodec()
 
-        chunks = _select_by_mask(self.shape_mask, (1,) * len(self.shape), self.internal_shape)
+        chunks = select_by_mask(self.shape_mask, (1,) * len(self.shape), self.internal_shape)
         self.array = zarr.open(
             self.store,
             mode="a",
@@ -86,7 +87,7 @@ class ZarrFileArray(StorageBase):
     def get_from_index(self, index: int) -> Any:
         """Return the data associated with the given linear index."""
         np_index = np.unravel_index(index, self.shape)
-        full_index = _select_by_mask(
+        full_index = select_by_mask(
             self.shape_mask,
             np_index,
             (slice(None),) * len(self.internal_shape),
@@ -141,12 +142,12 @@ class ZarrFileArray(StorageBase):
             raise NotImplementedError(msg)
 
         mask = self._mask[:]
-        slc = _select_by_mask(
+        slc = select_by_mask(
             self.shape_mask,
             (slice(None),) * len(self.shape),
             (None,) * len(self.internal_shape),  # Adds axes with size 1
         )
-        tile_shape = _select_by_mask(self.shape_mask, (1,) * len(self.shape), self.internal_shape)
+        tile_shape = select_by_mask(self.shape_mask, (1,) * len(self.shape), self.internal_shape)
         mask = np.tile(mask[slc], tile_shape)
 
         return np.ma.MaskedArray(self.array[:], mask=mask, dtype=object)
@@ -175,7 +176,7 @@ class ZarrFileArray(StorageBase):
                 if self.internal_shape:
                     value = np.asarray(value)  # in case it's a list
                     assert value.shape == self.internal_shape
-                    full_index = _select_by_mask(
+                    full_index = select_by_mask(
                         self.shape_mask,
                         external_index,
                         (slice(None),) * len(self.internal_shape),
@@ -190,7 +191,7 @@ class ZarrFileArray(StorageBase):
             value = np.asarray(value)  # in case it's a list
             assert value.shape == self.internal_shape
             assert len(key) == len(self.shape)
-            full_index = _select_by_mask(
+            full_index = select_by_mask(
                 self.shape_mask,
                 key,
                 (slice(None),) * len(self.internal_shape),
