+++ pipefunc/map/adaptive_scheduler.py
@@ -9,7 +9,7 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Any, NamedTuple
 
 from pipefunc._utils import at_least_tuple, requires
-from pipefunc.map._run import _func_kwargs, _load_file_arrays, _select_kwargs
+from pipefunc.map._run import _func_kwargs, _load_arrays, _select_kwargs
 from pipefunc.resources import Resources
 
 if TYPE_CHECKING:
@@ -200,7 +200,7 @@ def _eval_resources(
     run_info: RunInfo,
 ) -> Resources:
     kwargs = _func_kwargs(func, run_info, run_info.init_store())
-    _load_file_arrays(kwargs)
+    _load_arrays(kwargs)
     if index is not None:
         shape = run_info.shapes[func.output_name]
         shape_mask = run_info.shape_masks[func.output_name]
