+++ tests/map/storage/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for pipefunc.map.storage."""
